# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Whim 单组织 RBAC API",
    "version": "1.0.0",
    "description": "用户全局唯一，单主部门，多角色、多岗位。数据范围：1全部、2自定义部门、3本部门、4本部门及下级、5本人。岗位不会隐式授予角色。Long 超出 JavaScript 安全整数时序列化为字符串。"
  },
  "servers": [
    {
      "url": "http://localhost:8089"
    }
  ],
...
  "paths": {
...
    "/system/role": {
      "get": {
        "tags": [
          "角色"
        ],
        "summary": "查询系统角色。",
        "operationId": "listRoles",
        "description": "操作权限：`system:role:list`。功能授权在服务层校验当前有效账号及权限。",
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultRoleList"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "角色"
        ],
        "summary": "创建系统角色。",
        "operationId": "createRole",
        "description": "操作权限：`system:role:create`。功能授权在服务层校验当前有效账号及权限。",
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultId"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleSaveRequest"
              }
            }
          }
        }
      }
    },
    "/system/role/{roleId}": {
      "put": {
        "tags": [
          "角色"
        ],
        "summary": "修改系统角色。",
        "operationId": "updateRole",
        "description": "操作权限：`system:role:update`。功能授权在服务层校验当前有效账号及权限。",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "description": "目标主键 ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleSaveRequest"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "角色"
        ],
        "summary": "删除未被用户持有的普通角色。",
        "operationId": "deleteRole",
        "description": "操作权限：`system:role:delete`。功能授权在服务层校验当前有效账号及权限。",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "description": "目标主键 ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        }
      }
    },
    "/system/role/{roleId}/status": {
      "put": {
        "tags": [
          "角色"
        ],
        "summary": "修改系统角色状态。",
        "operationId": "setRoleStatus",
        "description": "操作权限：`system:role:status`。功能授权在服务层校验当前有效账号及权限。",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "description": "目标主键 ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusRequest"
              }
            }
          }
        }
      }
    },
    "/system/role/{roleId}/dataScope": {
      "get": {
        "tags": [
          "角色"
        ],
        "summary": "查询系统角色的数据范围。",
        "operationId": "getRoleDataScope",
        "description": "操作权限：`system:role:dataScope:list`。功能授权在服务层校验当前有效账号及权限。",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "description": "目标主键 ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultRoleDataScopeVO"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "角色"
        ],
        "summary": "覆盖系统角色的数据范围。",
        "operationId": "replaceRoleDataScope",
        "description": "操作权限：`system:role:dataScope:assign`。功能授权在服务层校验当前有效账号及权限。",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "description": "目标主键 ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleDataScopeRequest"
              }
            }
          }
        }
      }
    },
    "/system/role/{roleId}/permissions": {
      "get": {
        "tags": [
          "角色"
        ],
        "summary": "查询系统角色已分配的权限ID。",
        "operationId": "getRolePermissionIds",
        "description": "操作权限：`system:rolePermission:list`。功能授权在服务层校验当前有效账号及权限。",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "description": "目标主键 ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultIdList"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "角色"
        ],
        "summary": "覆盖系统角色的权限。",
        "operationId": "replaceRolePermissions",
        "description": "操作权限：`system:rolePermission:assign`。功能授权在服务层校验当前有效账号及权限。",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "description": "目标主键 ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolePermissionAssignRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "RoleSaveRequest": {
        "type": "object",
        "properties": {
          "roleName": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1
          },
          "roleCode": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1
          }
        },
        "required": [
          "roleName",
          "roleCode"
        ],
        "description": "系统角色保存参数。"
      },
      "RoleDataScopeRequest": {
        "type": "object",
        "properties": {
          "dataScope": {
            "type": "integer",
            "format": "int32"
          },
          "deptIds": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "type": "integer",
                  "format": "int64"
                },
                {
                  "type": "string",
                  "pattern": "^-?[0-9]+$"
                }
              ],
              "description": "Java Long；响应超出 JavaScript 安全整数范围时为字符串，请保持 ID 原始值。"
            },
            "uniqueItems": true
          }
        },
        "required": [
          "dataScope",
          "deptIds"
        ],
        "description": "系统角色数据范围配置参数。"
      },
...
      "RolePermissionAssignRequest": {
        "type": "object",
        "properties": {
          "permissionIds": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "type": "integer",
                  "format": "int64"
                },
                {
                  "type": "string",
                  "pattern": "^-?[0-9]+$"
                }
              ],
              "description": "Java Long；响应超出 JavaScript 安全整数范围时为字符串，请保持 ID 原始值。"
            },
            "uniqueItems": true
          }
        },
        "required": [
          "permissionIds"
        ],
        "description": "系统角色权限覆盖式分配参数。"
      },
...
      "StatusRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "status"
        ],
        "description": "全局权限状态修改参数。"
      },
      "ErrorResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "与 HTTP 状态码一致。"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "nullable": true,
            "oneOf": [
              {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "field": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            ]
          },
          "success": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "code",
          "message",
          "success"
        ]
      },
...
      "ResultVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string"
          },
          "data": {
            "nullable": true,
            "description": "无返回数据，值为 null。"
          },
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "success"
        ]
      },
      "ResultRoleList": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleData"
            }
          },
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "success"
        ]
      },
      "ResultId": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string"
          },
          "data": {
            "oneOf": [
              {
                "type": "integer",
                "format": "int64"
              },
              {
                "type": "string",
                "pattern": "^-?[0-9]+$"
              }
            ],
            "description": "Java Long；响应超出 JavaScript 安全整数范围时为字符串，请保持 ID 原始值。"
          },
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "success"
        ]
      },
      "ResultRoleDataScopeVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/RoleDataScopeData"
          },
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "success"
        ]
      },
      "ResultIdList": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "type": "integer",
                  "format": "int64"
                },
                {
                  "type": "string",
                  "pattern": "^-?[0-9]+$"
                }
              ],
              "description": "Java Long；响应超出 JavaScript 安全整数范围时为字符串，请保持 ID 原始值。"
            },
            "uniqueItems": true
          },
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "success"
        ]
      },
...
    }
  }
}